convert models page to hybrid server/client component
